Using Guided fragment for channel customization.
