+++ app/src/main/AndroidManifest.xml
@@ -6,6 +6,8 @@
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.RECORD_AUDIO"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="com.android.providers.tv.permission.READ_EPG_DATA" />
+    <uses-permission android:name="com.android.providers.tv.permission.WRITE_EPG_DATA" />
 
     <uses-feature
         android:name="android.hardware.touchscreen"
@@ -17,7 +19,7 @@
     <application
         android:allowBackup="true"
         android:icon="@mipmap/app_banner_sample_app"
-        android:label="@string/app_name"
+        android:banner="@string/app_name"
         android:supportsRtl="true"
         android:largeHeap="true"
         android:theme="@style/Theme.Example.LeanbackLauncher">
@@ -91,6 +93,7 @@
             android:exported="true"
             android:theme="@style/Theme.Example.LeanbackMusic"/>
         <activity
+            android:theme="@style/Theme.Example.LeanbackBrowse"
             android:name=".app.rows.DynamicVideoRowsActivity"
             android:label="@string/app_name">
             <intent-filter>
@@ -127,6 +130,10 @@
                     />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".app.rows.ChannelPublishActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Leanback.GuidedStep"/>
 
         <service android:exported="false" android:name=".app.media.MusicPlaybackService"/>
     </application>
