+++ app/src/main/java/android/support/v17/leanback/supportleanbackshowcase/app/rows/ChannelPublishActivity.java
@@ -1,0 +1,37 @@
+package android.support.v17.leanback.supportleanbackshowcase.app.rows;
+/*
+ * Copyright (c) 2017 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+import android.app.Activity;
+import android.os.Bundle;
+import android.support.v17.leanback.app.GuidedStepFragment;
+
+/**
+ * Activity that showcases different aspects of GuidedStepFragments.
+ */
+public class ChannelPublishActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (null == savedInstanceState) {
+            GuidedStepFragment.addAsRoot(this, new PublishChannelFragment(), android.R.id.content);
+        }
+    }
+
+}
+
