--- app/src/main/java/android/support/v17/leanback/supportleanbackshowcase/app/rows/VideoPlaybackFragment.java
@@ -1,115 +1,0 @@
-/*
- * Copyright (C) 2017 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package android.support.v17.leanback.supportleanbackshowcase.app.rows;
-
-import android.content.Context;
-import android.media.AudioManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v17.leanback.app.VideoFragment;
-import android.support.v17.leanback.app.VideoFragmentGlueHost;
-import android.support.v17.leanback.media.MediaPlayerAdapter;
-import android.support.v17.leanback.media.PlaybackGlue;
-import android.support.v17.leanback.supportleanbackshowcase.app.media.PlaybackSeekDiskDataProvider;
-import android.support.v17.leanback.supportleanbackshowcase.app.media.VideoMediaPlayerGlue;
-import android.util.Log;
-
-/**
- * A video playback fragment whose main task is to load and play videos
- */
-public class VideoPlaybackFragment extends VideoFragment {
-
-    private static final String URL = "https://storage.googleapis.com/android-tv/Sample videos/"
-            + "April Fool's 2013/Explore Treasure Mode with Google Maps.mp4";
-    public static final String TAG = "VideoConsumption";
-    private VideoMediaPlayerGlue<MediaPlayerAdapter> mMediaPlayerGlue;
-    final VideoFragmentGlueHost mHost = new VideoFragmentGlueHost(this);
-
-    static void playWhenReady(PlaybackGlue glue) {
-        if (glue.isPrepared()) {
-            glue.play();
-        } else {
-            glue.addPlayerCallback(new PlaybackGlue.PlayerCallback() {
-                @Override
-                public void onPreparedStateChanged(PlaybackGlue glue) {
-                    if (glue.isPrepared()) {
-                        glue.removePlayerCallback(this);
-                        glue.play();
-                    }
-                }
-            });
-        }
-    }
-
-    public static VideoPlaybackFragment newInstance(VideoContent selectedVideo) {
-        VideoPlaybackFragment videoPlaybackFragment = new VideoPlaybackFragment();
-        Bundle args = new Bundle(1);
-        args.putParcelable(VideoPlaybackActivity.VIDEO_CONTENT, selectedVideo);
-        videoPlaybackFragment.setArguments(args);
-        return videoPlaybackFragment;
-    }
-
-    AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener
-            = new AudioManager.OnAudioFocusChangeListener() {
-        @Override
-        public void onAudioFocusChange(int state) {
-        }
-    };
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        mMediaPlayerGlue = new VideoMediaPlayerGlue(getActivity(),
-                new MediaPlayerAdapter(getActivity()));
-        mMediaPlayerGlue.setHost(mHost);
-
-        AudioManager audioManager = (AudioManager) getActivity()
-                .getSystemService(Context.AUDIO_SERVICE);
-        if (audioManager.requestAudioFocus(mOnAudioFocusChangeListener, AudioManager.STREAM_MUSIC,
-                AudioManager.AUDIOFOCUS_GAIN) != AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-            Log.w(TAG, "video player cannot obtain audio focus!");
-        }
-
-        Bundle args = getArguments();
-        VideoContent video = args.getParcelable(VideoPlaybackActivity.VIDEO_CONTENT);
-        if (video != null) {
-            mMediaPlayerGlue.setTitle(video.getTitle());
-            mMediaPlayerGlue.setSubtitle(video.getDescription());
-            mMediaPlayerGlue.getPlayerAdapter().setDataSource(
-                    Uri.parse(video.getVideoUrl()));
-        } else {
-            // when videoContent has not been set, play the default video
-            mMediaPlayerGlue.setTitle("Diving with Sharks");
-            mMediaPlayerGlue.setSubtitle("A Googler");
-            mMediaPlayerGlue.getPlayerAdapter().setDataSource(Uri.parse(URL));
-        }
-        PlaybackSeekDiskDataProvider.setDemoSeekProvider(mMediaPlayerGlue);
-        playWhenReady(mMediaPlayerGlue);
-        setBackgroundType(BG_LIGHT);
-    }
-
-    @Override
-    public void onPause() {
-        if (mMediaPlayerGlue != null) {
-            mMediaPlayerGlue.pause();
-        }
-        super.onPause();
-    }
-
-}
-
