+++ app/src/main/res/values/strings.xml
@@ -20,6 +20,7 @@
     <!-- Error messages -->
     <string name="implement_search">Implement your own in-app search.</string>
     <string name="card_examples_title">Card Examples</string>
+    <string name="channel_publish_title">Channel Publish</string>
     <string name="detail_view_title">DetailView Example</string>
     <string name="action_cicked">Action clicked. Implement your own handler.</string>
     <string name="grid_example_title">Image Grid Example</string>
@@ -75,4 +76,17 @@
     <string name="remove_program" translatable="false">Remove Program</string>
     <string name="another_option" translatable="false">Another Option</string>
     <string name="another_option_message" translatable="false">Another program action</string>
+
+    <!--channel publish customization-->
+    <string name="setting_page_title">Publish Channels</string>
+    <string name="setting_page_description">Toggle the check box</string>
+    <string name="setting_page_breadcrumb">Channel Customization</string>
+    <string name="channel_toggle_action_description">Toggle the checkbox</string>
+    <string name="channel_toggle_action_title">Choose channel</string>
+    <string name="channel_toggle_action_finish_desc">Click when you finished</string>
+    <string name="channel_toggle_action_finish_title">Publish</string>
+    <string name="finish_action_title">Done</string>
+    <string name="finish_action_description">Channel will be shown in main screen</string>
+    <string name="finish_action_clicked">Channel Customization Finished</string>
+    <string name="channel_removed_from_home_screen">Channel is removed from home screen</string>
 </resources>
